restore legacy account to master branch version
