--- contracts/warp-legacy-account/src/execute/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod ibc;
-pub(crate) mod withdraw;
