+++ contracts/warp-legacy-account/src/lib.rs
@@ -1,7 +1,5 @@
 pub mod contract;
 mod error;
-mod execute;
-mod query;
 pub mod state;
 
 #[cfg(test)]
