--- contracts/warp-legacy-account/src/query/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod account;
