+++ packages/job-account/src/lib.rs
@@ -12,11 +12,6 @@ pub struct Config {
 
     // Address of account tracker contract
     pub job_account_tracker_addr: Addr,
-    // // Address of current warp account contract
-    // pub account_addr: Addr,
-
-    // // If occupied, occupied_by_job_id is the job id of the pending job that is using this sub account
-    // pub occupied_by_job_id: Option<Uint64>,
 }
 
 #[cw_serde]
@@ -31,8 +26,7 @@ pub struct InstantiateMsg {
     // But store it enables us the other way around
     pub job_account_tracker_addr: String,
 
-    // Only required when we are instantiate a main account
-    // Since we always want to fund sub account, so we will pass this value around and send it to sub account during instantiation in create main account's reply
+    // Native funds
     pub native_funds: Vec<NativeCoin>,
     // CW20 or CW721 funds, will be transferred to account in reply of account instantiation
     pub cw_funds: Vec<CwFund>,
