+++ packages/legacy-account/src/lib.rs
@@ -1,43 +1,19 @@
 use controller::account::{AssetInfo, CwFund};
-use cosmwasm_schema::{cw_serde, QueryResponses};
-use cosmwasm_std::{Addr, Coin as NativeCoin, CosmosMsg, Uint64};
+use cosmwasm_schema::cw_serde;
+use cosmwasm_std::{Addr, CosmosMsg};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 #[cw_serde]
 pub struct Config {
     pub owner: Addr,
-    // Address of warp controller contract
-    pub creator_addr: Addr,
-
-    // Address of account tracker contract
-    pub job_account_tracker_addr: Addr,
-    // // Address of current warp account contract
-    // pub account_addr: Addr,
-
-    // // If occupied, occupied_by_job_id is the job id of the pending job that is using this sub account
-    // pub occupied_by_job_id: Option<Uint64>,
+    pub warp_addr: Addr,
 }
 
 #[cw_serde]
 pub struct InstantiateMsg {
-    // User who owns this account
     pub owner: String,
-    // ID of the job that is created along with the account
-    pub job_id: Uint64,
-
-    // Account tracker tracks all accounts owned by user
-    // Store it inside account for easier lookup, though most of time we only lookup account from account tracker
-    // But store it enables us the other way around
-    pub job_account_tracker_addr: String,
-
-    // Only required when we are instantiate a main account
-    // Since we always want to fund sub account, so we will pass this value around and send it to sub account during instantiation in create main account's reply
-    pub native_funds: Vec<NativeCoin>,
-    // CW20 or CW721 funds, will be transferred to account in reply of account instantiation
-    pub cw_funds: Vec<CwFund>,
-    // List of cosmos msgs to execute after instantiating the account
-    pub msgs: Vec<CosmosMsg>,
+    pub funds: Option<Vec<CwFund>>,
 }
 
 #[cw_serde]
@@ -110,19 +86,9 @@ pub struct WithdrawAssetsMsg {
 #[cw_serde]
 pub struct ExecuteWasmMsg {}
 
-#[derive(QueryResponses)]
 #[cw_serde]
 pub enum QueryMsg {
-    #[returns(ConfigResponse)]
-    QueryConfig(QueryConfigMsg),
-}
-
-#[cw_serde]
-pub struct QueryConfigMsg {}
-
-#[cw_serde]
-pub struct ConfigResponse {
-    pub config: Config,
+    Config,
 }
 
 #[cw_serde]
